1.code/requirements.txt
Flask==2.3.3
flask-sqlalchemy==3.0.5
flask-restx==1.2.0
pymysql==1.1.0
python-dotenv==1.0.0
requests==2.31.0
boto3==1.28.85
botocore==1.31.85
APScheduler==3.10.4
cryptography==41.0.7